refactor: rename tinybird-api.ts to api.ts
